Refactor #137 Rename the 'controller.js' file to 'employee.js'
